good stuff

+++ controllers/productController.js
@@ -13,7 +13,7 @@ export const createProduct = async (req, res) => {
 
     res.status(201).json(newProduct);
   } catch (error) {
-    res.status(500).json({ message: 'Error creating product', error });
+    res.status(500).json({ message: 'Error 500. Error creating product', error });
   }
 };
 
@@ -22,7 +22,7 @@ export const getProducts = async (req, res) => {
     const products = await Product.find();
     res.status(200).json(products);
   } catch (error) {
-    res.status(500).json({ message: 'Error retrieving products', error });
+    res.status(500).json({ message: 'Error 500. Error retrieving products', error });
   }
 };
 
@@ -31,19 +31,25 @@ export const updateProduct = async (req, res) => {
   const { productName, cost, amountAvailable } = req.body;
 
   try {
-    const updatedProduct = await Product.findByIdAndUpdate(productId, {
+    const updatedProduct = await Product.findById(productId);
+
+    if (!updatedProduct) {
+      return res.status(404).json({ message: 'Error 404. Product not found' });
+    }
+
+    if (updatedProduct.sellerId.toString() !== req.user.id) {
+      return res.status(403).json({ message: 'Error 403. You are not allowed to update other sellers products.' });
+    }
+
+    await Product.findByIdAndUpdate(productId, {
       productName,
       cost,
       amountAvailable
     }, { new: true });
 
-    if (!updatedProduct) {
-      return res.status(404).json({ message: 'Product not found' });
-    }
-
     res.status(200).json(updatedProduct);
   } catch (error) {
-    res.status(500).json({ message: 'Error updating product', error });
+    res.status(500).json({ message: 'Error 500. Error updating product', error });
   }
 };
 
@@ -51,14 +57,19 @@ export const deleteProduct = async (req, res) => {
   const { productId } = req.params;
 
   try {
-    const deletedProduct = await Product.findByIdAndDelete(productId);
+    const product = await Product.findById(productId);
 
-    if (!deletedProduct) {
-      return res.status(404).json({ message: 'Product not found' });
+    if (!product) {
+      return res.status(404).json({ message: 'Error 404. Product not found' });
     }
 
+    if (product.sellerId.toString() !== req.user.id) {
+      return res.status(403).json({ message: 'Error 403. You are not allowed to delete other sellers products.' });
+    }
+
+    await Product.findByIdAndDelete(productId);
     res.status(200).json({ message: 'Product deleted successfully' });
   } catch (error) {
-    res.status(500).json({ message: 'Error deleting product', error });
+    res.status(500).json({ message: 'Error 500. Error deleting product', error });
   }
 };
